requirements.txt
pytest==7.1.2
sqlmodel==0.0.6
fastapi==0.78.0
uvicorn==0.17.6
python-multipart==0.0.5
python-dotenv==0.20.0
